creation fichier affichageResultat + modification de l'import du fichier
