+++ templates/importJSON.php
@@ -1,20 +1,20 @@
 <?php
-require_once 'quizz.php';
+session_start();
+require_once 'affichageQuizz.php';
 use function templates\affichageQuizz;
 
-if ($_SERVER["REQUEST_METHOD"] === "POST") {
-    if (isset($_FILES["quizz"])) {
+if ($_SERVER["REQUEST_METHOD"] === "POST" && isset($_POST['importQuizz'])) {
+    if (isset($_FILES["quizzJSON"])) {
         $dossier = "../Data/import/"; 
-        $nomFichier = basename($_FILES["quizz"]["name"]);
+        $nomFichier = basename($_FILES["quizzJSON"]["name"]);
         $fic = $dossier . $nomFichier;
-
-        if (move_uploaded_file($_FILES["quizz"]["tmp_name"], $fic)) {
+        if (move_uploaded_file($_FILES["quizzJSON"]["tmp_name"], $fic)) {
             affichageQuizz($fic);
         } else {
             echo "Erreur lors du téléchargement.";
         }
     } else {
-        echo "Aucun fichier.";
+        echo "Aucun fichier sélectionné.";
     }
 }
 ?>
